ZMLP 355 Tracking deleted assets

See merge request zorroa-zvi/zmlp!1224

+++ services/archivist/archivist/src/main/kotlin/com/zorroa/archivist/service/AssetService.kt
@@ -442,7 +442,13 @@ class AssetServiceImpl : AssetService {
 
     override fun batchDelete(ids: Set<String>): BatchDeleteAssetResponse {
 
+        val maximumBatchSize = properties.getInt("archivist.assets.deletion-max-batch-size")
+        if (ids.size > maximumBatchSize) {
+            throw IllegalArgumentException("Maximum allowed size exceeded. Maximum batch size for delete: $maximumBatchSize")
+        }
+
         val rest = indexRoutingService.getProjectRestClient()
+        var deletedAssets = getAll(ids).map { it.id to it }.toMap()
         val query = QueryBuilders.termsQuery("_id", ids)
         val rsp = rest.client.deleteByQuery(
             DeleteByQueryRequest(rest.route.indexName)
@@ -461,11 +467,14 @@ class AssetServiceImpl : AssetService {
             )
         }
 
+        val projectQuotaCounters = ProjectQuotaCounters()
         for (removed in removed) {
+            projectQuotaCounters.countForDeletion(deletedAssets.getValue(removed))
             logger.event(
                 LogObject.ASSET, LogAction.DELETE, mapOf("assetId" to removed)
             )
         }
+        projectService.incrementQuotaCounters(projectQuotaCounters)
 
         // Background removal of files into a co-routine.
         GlobalScope.launch(Dispatchers.IO + CoroutineAuthentication(SecurityContextHolder.getContext())) {
